LPS-46205 Rename package and remove 'model' package
