+++ portal-service/src/com/liferay/portal/kernel/expression/ExpressionEvaluator.java
@@ -1,7 +1,6 @@
 
-package com.liferay.portal.kernel.expressionevaluator;
+package com.liferay.portal.kernel.expression;
 
-import com.liferay.portal.kernel.expressionevaluator.model.VariableDependencies;
 
 import java.lang.reflect.InvocationTargetException;
