+++ portal-service/src/com/liferay/portal/kernel/expression/Functions.java
@@ -1,5 +1,5 @@
 
-package com.liferay.portal.kernel.expressionevaluator;
+package com.liferay.portal.kernel.expression;
 
 /**
  * @author Miguel Angelo Caldas Gallindo
