+++ portal-service/test/unit/com/liferay/portal/kernel/expressionevaluator/DependenciesTest.java
@@ -1,7 +1,9 @@
 
 package com.liferay.portal.kernel.expressionevaluator;
 
-import com.liferay.portal.kernel.expressionevaluator.model.VariableDependencies;
+import com.liferay.portal.kernel.expression.ExpressionEvaluator;
+import com.liferay.portal.kernel.expression.ExpressionEvaluatorImpl;
+import com.liferay.portal.kernel.expression.VariableDependencies;
 
 import java.util.Map;
